requirements.txt
fastapi 
uvicorn
python-dotenv
fpdf2
thefuzz
python-multipart
pdf2image
paddleocr
sentence-transformers
boto3